# go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "统一编码格式-UTF8\n",
        "title": "验证系统 API",
        "contact": {},
        "license": {
            "name": "易轩软件 http://www.99v66.com",
            "url": "http://www.99v66.com"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:52777",
    "basePath": "/v1",
    "paths": {
...
        "/power/charge": {
            "post": {
                "description": "详细描述",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户操作"
                ],
                "summary": "充值",
                "parameters": [
                    {
                        "type": "string",
                        "description": "账号",
                        "name": "user",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "卡密",
                        "name": "cdk",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "应用ID",
                        "name": "aid",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"充值成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/power/check": {
            "post": {
                "description": "数据已加密",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "验证相关"
                ],
                "summary": "心跳[数据已加密]",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录返回的token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "机器码",
                        "name": "dev",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"效验成功\",\"data\":{\"a\":{\"aid\":3,\"name\":\"JD注册\",\"mold\":2,\"login\":3,\"bin\":1,\"msg\":\"关键参数222\",\"state\":1,\"time\":1625069421},\"p\":{\"pid\":3,\"uid\":1,\"aid\":3,\"kid\":0,\"day\":1625183400,\"point\":90,\"ip\":\"192.168.1.7\",\"code\":\"ca5df0fdae677510f937c256a7f5\",\"state\":1,\"time\":1625084599}}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/power/edit": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "权限管理"
                ],
                "summary": "编辑权限信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录返回的token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "uid",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "应用ID",
                        "name": "aid",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "1=可用 2=不可用",
                        "name": "state",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "到期时间的十位时间戳",
                        "name": "day",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "剩余点数",
                        "name": "point",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "绑定IP",
                        "name": "ip",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "绑定机器",
                        "name": "code",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"修改成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/power/getAll": {
            "get": {
                "description": "详细描述",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "权限管理"
                ],
                "summary": "获取权限列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录返回的token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "页数",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "获取的数量",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "应用ID",
                        "name": "aid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "状态1=可用 2=不可用",
                        "name": "state",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "uid",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"查询成功\",\"count\":1,\"data\":[{\"uid\":1001,\"aid\":1002,\"kid\":0,\"day\":1625191303,\"point\":200,\"ip\":\"\",\"code\":\"\",\"state\":1,\"time\":1625018503}]}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/power/login": {
            "post": {
                "description": "数据已加密",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "验证相关"
                ],
                "summary": "用户登录[数据已加密]",
                "parameters": [
                    {
                        "type": "string",
                        "description": "账号",
                        "name": "user",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "pwd",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "应用ID",
                        "name": "aid",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "机器码",
                        "name": "dev",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"登陆成功\",\"data\":{\"token\":\"111...\"}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/power/sub": {
            "post": {
                "description": "数据已加密",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "验证相关"
                ],
                "summary": "扣点[数据已加密]",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录返回的token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "扣除点数不可为负",
                        "name": "point",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"扣除成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    }
}